requirements.txt
langchain==0.0.138
nltk==3.7
openai==0.27.2
pinecone_client==2.2.1
python-dotenv==1.0.0
PyYAML==6.0
flask==2.3.2
gunicorn==20.1.0
